Replace name-based syntax with extension-based syntax for tagged paths
The previous name-based syntax looked like,
`TAG1-TAG2-_NAME`.
The new extension-based syntax looks like,
`TAG1-TAG2.EXT`.
The new syntax can result in cleaner and more compact paths,
especially when tags are derived from names.

Tagged paths now support ignored tags.
Ignored tags are not treated as tags,
but they are allowed anywhere tags are.
This is needed to differentiate paths with the same tags and extension.
Ignored tags start with `_`.

Most of the work on this patch is related to resolving ambiguity
caused by paths with the same tags and extensions.
Previously,
even path needed a unique name,
which prevented ambiguity.